Fix rpm packaging issue with shebang calculation
When this module is installed into virtualenv and it is packed into rpm
package, an issue with executable files arises.
The rpm building software treats executable files as ones containing
shebang lines, and the first word is treated as an interpretator, which,
it its turn goes to rpm dependency list.
With first word `#cython` in _types.pyx or `import` in .py files the
resulting dependency is really wrong.

Given that this files are not really executables with correct shebang,
remove +x from them.